% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_annotations.R
\name{annotation_2_json}
\alias{annotation_2_json}
\title{annotation to json}
\usage{
annotation_2_json(annotation_obj, json_file = NULL)
}
\arguments{
\item{annotation_obj}{the annotation object}

\item{json_file}{the file to save it to}
}
\value{
the json string (invisibly)
}
\description{
Given a `categoryCompare2` annotation object, generate a JSON representation
that can be used with the command line executable
}
