% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{annotation_combinations,cc_graph-method}
\alias{annotation_combinations,cc_graph-method}
\title{unique combinations}
\usage{
\S4method{annotation_combinations}{cc_graph}(object)
}
\arguments{
\item{object}{the \linkS4class{cc_graph} to work on}
}
\value{
node_assignment
}
\description{
determine the unique combinations of annotations that exist in the
significant matrix of the \linkS4class{cc_graph} and assign each node in the graph
to a group.
}
