% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\docType{methods}
\name{annotation_combinations,significant_annotations-method}
\alias{annotation_combinations,significant_annotations-method}
\title{unique combinations}
\usage{
\S4method{annotation_combinations}{significant_annotations}(object)
}
\arguments{
\item{object}{the \linkS4class{combined_statistics} to work on}
}
\value{
node_assignment
}
\description{
determine the unique combinations of annotations that exist in the
significant matrix of the \linkS4class{combined_statistics} and assign each
annotation to a group.
}

