% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/graph_visualization.R
\name{annotation_combinations}
\alias{annotation_combinations}
\alias{annotation_combinations,cc_graph-method}
\alias{annotation_combinations,significant_annotations-method}
\title{unique annotation combinations}
\usage{
annotation_combinations(object)

\S4method{annotation_combinations}{cc_graph}(object)

\S4method{annotation_combinations}{significant_annotations}(object)
}
\arguments{
\item{object}{the \code{\link{combined_statistics}} to work on}
}
\value{
node_assignment

node_assignment
}
\description{
determine the unique combinations of annotations that exist in the
significant matrix of the \code{\link{cc_graph}} and assign each node in the graph
to a group.

determine the unique combinations of annotations that exist in the
significant matrix of the \code{\link{combined_statistics}} and assign each
annotation to a group.
}
