% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_classes.R
\name{annotation}
\alias{annotation}
\title{annotation constructor}
\usage{
annotation(annotation_features, type = NULL, description = character(0),
  links = character(0))
}
\arguments{
\item{annotation_features}{list of annotation to feature relationships}

\item{type}{a one word short description of the "type" of annotation}

\item{description}{character vector providing descriptive text about the annotation}

\item{links}{character vector defining html links for each annotation (may be empty)}

\item{counts}{numeric vector of how many features are in each annotation}
}
\description{
Does sensical checks when creating an \linkS4class{annotation} object.
}
\details{
See the \linkS4class{annotation}, each slot is a parameter.
}

