% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{assign_communities}
\alias{assign_communities}
\title{assign communities}
\usage{
assign_communities(in_graph)
}
\arguments{
\item{in_graph}{the \code{cc_graph} object to use}
}
\value{
list
}
\description{
given a \code{cc_graph}, find communities of nodes based on their connectivity
and weights.
}

