% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_enrichment.R
\name{binomial_basic}
\alias{binomial_basic}
\title{do binomial test}
\usage{
binomial_basic(
  positive_cases,
  total_cases,
  p_expected = 0.5,
  direction = "two.sided",
  conf_level = 0.95
)
}
\arguments{
\item{positive_cases}{number of positive instances}

\item{total_cases}{total number of cases observed}

\item{p_expected}{what is the expected probability}

\item{direction}{which direction is the test}

\item{conf_level}{confidence level for the confidence interval}
}
\value{
list
}
\description{
does a binomial test
}
