% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_enrichment.R
\name{binomial_feature_enrichment}
\alias{binomial_feature_enrichment}
\title{do binomial testing}
\usage{
binomial_feature_enrichment(
  binomial_features,
  p_expected = 0.5,
  direction = "two.sided",
  p_adjust = "none",
  conf_level = 0.95
)
}
\arguments{
\item{binomial_features}{a binomial_features object}

\item{p_expected}{the expected probability (default 0.5)}

\item{direction}{which direction to do the enrichment (two.sided, less, greater)}

\item{p_adjust}{how to correct the p-values (default is "none")}

\item{conf_level}{the confidence level for the confidence interval (default is 0.95)}
}
\value{
enriched_result
}
\description{
do binomial testing
}
