% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_enrichment.R
\docType{class}
\name{binomial_features-class}
\alias{binomial_features-class}
\title{binomial feature class}
\description{
class to hold features undergoing binomail statistical testing
}
\section{Slots}{

\describe{
\item{\code{positivefc}}{the features with positive fold-changes}

\item{\code{negativefc}}{the features with negative fold-changes}

\item{\code{annotation}}{annotation object}
}}

