% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_enrichment.R
\name{binomial_features}
\alias{binomial_features}
\title{binomial_features constructor}
\usage{
binomial_features(positivefc, negativefc, annotation)
}
\arguments{
\item{positivefc}{the positively changed features}

\item{negativefc}{the negatively changed features}

\item{annotation}{the annotation object}
}
\value{
binomial_features-class
}
\description{
Create the binomial_features-class object.
}
