% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{binomial_result-class}
\alias{binomial_result-class}
\alias{binomial_result}
\title{the binomial results class}
\description{
the binomial results class
}
\section{Slots}{

\describe{
\item{\code{positivefc}}{the positive log-fold-changed genes, a vector of class ANY}

\item{\code{negativefc}}{the negative log-fold-changed genes}

\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \code{\link{statistical_results}} object}
}}

