% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoryCompare2.R
\name{categoryCompare2}
\alias{categoryCompare2}
\title{categoryCompare2: A package for comparing enrichment results from multiple experiments}
\description{
The categoryCompare2 package provides functions for simple enrichment and
  and comparison of those enrichment results.
}
