% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/combine_enrichments.R
\name{combine_annotations}
\alias{combine_annotations}
\alias{combine_annotations,list-method}
\title{combine annotations}
\usage{
combine_annotations(annotation_list)

\S4method{combine_annotations}{list}(annotation_list)
}
\arguments{
\item{annotation_list}{one or more \code{\link{annotation}}}
}
\value{
\code{\link{annotation}}
}
\description{
Takes multiple \code{\link{annotation}} objects and combines them so that there
is a consistent sole set for creating the \code{\link{cc_graph}} and providing
other information about each annotation entry.
}
