% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/combine_enrichments.R
\name{combine_enrichments}
\alias{combine_enrichments}
\alias{combine_enrichments,enriched_result-method}
\alias{combine_enrichments,list-method}
\title{combine enrichments}
\usage{
combine_enrichments(...)

\S4method{combine_enrichments}{enriched_result}(...)

\S4method{combine_enrichments}{list}(...)
}
\arguments{
\item{...}{list of enriched_result}
}
\value{
\code{\link{combined_enrichment}}
}
\description{
This is one of the primary workhorse functions behind \pkg{categoryCompare2}.
The primary function of \code{categoryCompare} is to enable \emph{comparisons}
of different enrichment analyses. To facilitate that, we must first 
\strong{combine} one (really, we can do this with a single) or more 
\code{\link{enriched_result}}.
}
