% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{combined_enrichment-class}
\alias{combined_enrichment-class}
\alias{combined_enrichment}
\title{combined enrichments}
\description{
The \code{combined_enrichment} class holds the results of combining several 
\code{\link{enriched_result}}s together, which includes the original 
\code{\link{enriched_result}}s, as well as the \code{\link{cc_graph}}
and combined \code{\link{annotation}} objects.
}
\section{Slots}{

\describe{
\item{\code{enriched}}{list of enriched objects}

\item{\code{annotation}}{\code{\link{annotation}} where the annotation_features
have been combined across the \code{\link{enriched_result}}}

\item{\code{statistics}}{\code{\link{combined_statistics}} of both}
}}

