% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{count_go_children}
\alias{count_go_children}
\title{GO children}
\usage{
count_go_children(go_terms, which_go = c("BP", "MF", "CC"))
}
\arguments{
\item{go_terms}{the terms to do counting on}

\item{which_go}{which Gene Ontology should be used?}
}
\value{
numeric
}
\description{
counts all of the children for particular set of GO terms.
}

