% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{csv_annotation_table}
\alias{csv_annotation_table}
\title{print table csv}
\usage{
csv_annotation_table(annotation_gene_table, out_file = NULL)
}
\arguments{
\item{annotation_gene_table}{list of tables}

\item{out_file}{the file to write to}
}
\description{
print the annotation gene table to a CSV file
}
