% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{enriched_result-class}
\alias{enriched_result-class}
\alias{enriched_result}
\title{the enriched results class}
\description{
the enriched results class
}
\section{Slots}{

\describe{
\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \code{\link{statistical_results}} object}
}}

