% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{enriched_result}
\alias{enriched_result}
\title{the enriched results class}
\usage{
enriched_result(features, universe, annotation, statistics)

enriched_result(features, universe, annotation, statistics)
}
\arguments{
\item{features}{the features that were differentially expressed (see details)}

\item{universe}{all of the features that were measured}

\item{annotation}{an \code{\link{annotation}} object}

\item{statistics}{a \code{\link{statistical_results}} object}
}
\value{
enriched_result
}
\description{
given all the slots for an \code{\link{enriched_result}}, checks that all
the data is self-consistent, and creates the \code{\link{enriched_result}} object.
}
\section{Slots}{

\describe{
\item{\code{features}}{the "features" of interest, a vector of class ANY}

\item{\code{universe}}{all of the "features" in the background}

\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \code{\link{statistical_results}} object}
}}

