% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_classes.R
\name{enriched_result}
\alias{enriched_result}
\title{creates enriched result}
\usage{
enriched_result(features, universe, annotation, statistics)
}
\arguments{
\item{features}{the features that were differentially expressed (see details)}

\item{universe}{all of the features that were measured}

\item{annotation}{an \linkS4class{annotation} object}

\item{statistics}{a \linkS4class{statistical_results} object}
}
\value{
enriched_result
}
\description{
given all the slots for an \linkS4class{enriched_result}, checks that all
the data is self-consistent, and creates the \code{enriched_result} object.
}

