% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{extract_statistics,enriched_result-method}
\alias{extract_statistics,enriched_result-method}
\title{extract statistics}
\usage{
\S4method{extract_statistics}{enriched_result}(in_results)
}
\arguments{
\item{in_results}{the \code{\link{enriched_result}} object}
}
\value{
data.frame
}
\description{
extract single set of statistics from a \code{\link{enriched_result}} object and
create a \code{data.frame}.
}
