% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/combine_enrichments.R
\name{extract_statistics}
\alias{extract_statistics}
\alias{extract_statistics,statistical_results-method}
\title{get statistics}
\usage{
extract_statistics(in_results)

\S4method{extract_statistics}{statistical_results}(in_results)
}
\arguments{
\item{in_results}{the \code{\link{statistical_results}} object}
}
\value{
data.frame
}
\description{
extract all statistics for a \code{\link{statistical_results}} object. These
can then be combined into a \code{data.frame} that can be returned or used
to annotate the graph of annotations.
}
