% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/combine_enrichments.R
\name{generate_annotation_graph}
\alias{generate_annotation_graph}
\alias{generate_annotation_graph,combined_enrichment-method}
\title{generate annotation graph}
\usage{
generate_annotation_graph(
  comb_enrichment,
  annotation_similarity = "combined",
  low_cut = 5,
  hi_cut = 500
)

\S4method{generate_annotation_graph}{combined_enrichment}(
  comb_enrichment,
  annotation_similarity = "combined",
  low_cut = 5,
  hi_cut = 500
)
}
\arguments{
\item{comb_enrichment}{the combined_enrichment object}

\item{annotation_similarity}{which similarity measure to use}

\item{low_cut}{keep only those annotations in the graph with at least this many annotated features}

\item{hi_cut}{keep only those annotations with less than this many annotated features}
}
\value{
\code{\link{cc_graph}}
}
\description{
given a \code{\link{combined_enrichment}}, generate the annotation similarity graph
}
