% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/combine_enrichments.R
\name{generate_table}
\alias{generate_table}
\alias{generate_table,combined_enrichment-method}
\title{generate statistical table}
\usage{
generate_table(comb_enrichment, entries = "signficant", link_type = "explicit")

\S4method{generate_table}{combined_enrichment}(
  comb_enrichment,
  entries = "significant",
  link_type = "explicit"
)
}
\arguments{
\item{comb_enrichment}{the \code{\link{combined_enrichment}} object}

\item{entries}{which entries to return, "significant" or "all"}

\item{link_type}{should their be an "explicit" link (see details)}
}
\value{
data.frame
}
\description{
given a \code{\link{combined_enrichment}} object, get out the data.frame
either for investigation or to add data to the \code{\link{cc_graph}}.
}
\details{
the \code{link_type} controls whether to create an "explicit" link
  that is actually a column in the data.frame, or create an "implicit" html link
  that is part of the \code{@name} column in the returned data.frame. Useful
  if you are embedding the data.frame in an html report.
}
