% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_annotations.R
\name{get_db_annotation}
\alias{get_db_annotation}
\title{orgdb annotations}
\usage{
get_db_annotation(
  orgdb = "org.Hs.eg.db",
  features = NULL,
  feature_type = "ENTREZID",
  annotation_type = "GO"
)
}
\arguments{
\item{orgdb}{the name of the org.*.db object}

\item{feature_type}{which type of IDs to map (see details)}

\item{annotation_type}{the type of annotation to grab (see details)}
}
\value{
annotation object
}
\description{
Generate an annotation object for genes based on an "org.*.db" object, and pulling
information from it.
}
\details{
This function generates a \code{categoryCompare2} annotation object
from a Bioconductor "org.*.db" object. Even though different gene identifiers can
be used, almost all of the mappings are via ENTREZID.

The set of feature or gene keys that can be used to create the annotations include:
\itemize{
\item ENTREZID: ENTREZ gene ids
\item ACCNUM: genbank accession numbers
\item SYMBOL: gene symbols, eg ABCA1
\item GENENAME: gene names, eg "ATP binding cassette subfamily A member 1"
\item ENSEMBL: the ensembl gene ids (all start with ENSG...)
\item ENSEMBLPROT: ensembl protein ids (ENSP...)
\item ENSEMBLTRANS: ensemlb transcript ids (ENST...)
\item REFSEQ: reference sequence IDs, NM, NP, NR, XP, etc
\item UNIGENE: gene ids from UNIPROT eg Hs.88556
\item UNIPROT: protein ids from UNIPROT eg P80404
}

The set of annotations that can be mapped to features include:
\itemize{
\item GO: annotations from gene ontology
\item PATH: KEGG Pathway identifiers (not updated since 2011!)
\item CHRLOC: location on the chromosome
\item OMIM: mendelian inheritance in man identifiers
\item PMID: pubmed identifiers
\item PROSITE
\item PFAM: protein family identifiers
\item IPI: protein-protein interactions
}

For GO annotations, it is also possible to pass \code{GO} to use all 3 sub-ontologies simultaneously,
or any combination of \code{BP}, \code{MF}, and \code{CC}.
}
