% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/significant_results.R
\name{get_significant_annotations}
\alias{get_significant_annotations}
\alias{get_significant_annotations,statistical_results-method}
\alias{get_significant_annotations,combined_enrichment-method}
\title{get significant annotations}
\usage{
get_significant_annotations(in_results, ...)

\S4method{get_significant_annotations}{statistical_results}(in_results, ...)

\S4method{get_significant_annotations}{combined_enrichment}(in_results, ...)
}
\arguments{
\item{in_results}{a \code{\link{combined_enrichment}} object}

\item{...}{conditional expressions}
}
\value{
vector of significant annotation_id's

\code{\link{combined_enrichment}} object
}
\description{
given a \code{\link{statistical_results}} object and some conditional expressions,
return the significant annotations

In the case where we have a \code{\link{combined_enrichment}} and we want
to get all of the significant annotations from each of them, and put them
together so we can start doing real meta-analysis.
}
\details{
Note that this function returns the original \code{\link{combined_enrichment}} object with a modified
\code{\link{combined_statistics}} slot where the significant annotations have been added in.
}
\examples{

test_stat <- new("statistical_results",
                 annotation_id = c("a1", "a2", "a3"),
                 statistic_data = list(pvalues = c(a1 = 0.01, a2 = 0.5, a3 = 0.0001),
                   counts = c(a1 = 5, a2 = 10, a3 = 1),
                   odds = c(a1 = 20, a2 = 100, a3 = 0)))
get_significant_annotations(test_stat, pvalues < 0.05)
get_significant_annotations(test_stat, odds > 10)
get_significant_annotations(test_stat, pvalues < 0.05, counts >= 1)

}
