% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_results_calls.R
\name{get_significant_annotations_calls}
\alias{get_significant_annotations_calls}
\title{get significant annotations calls}
\usage{
get_significant_annotations_calls(in_results, queries)
}
\arguments{
\item{in_results}{a \linkS4class{statistical_results} object}

\item{queries}{a list of queries that can form a call object}
}
\value{
vector of significant annotation_id's
}
\description{
In the case where we have a \linkS4class{statistical_results} and we want
to get all of the significant annotations from it
}
