% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_annotations.R
\name{gocats_to_annotation}
\alias{gocats_to_annotation}
\title{annotation from GOcats}
\usage{
gocats_to_annotation(json_file, feature_type = NULL, annotation_type = NULL)
}
\arguments{
\item{json_file}{the json file to use}

\item{feature_type}{the type of features}

\item{annotation_type}{the type of annotations}
}
\value{
an annotation object
}
\description{
Given a JSON file of features to annotations from GOcats, reverse to turn it into
annotations to features, and optionally add some meta-information about them.
}
