% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{graph_to_visnetwork}
\alias{graph_to_visnetwork}
\title{cc_graph to visnetwork}
\usage{
graph_to_visnetwork(
  in_graph,
  in_assign,
  node_communities = NULL,
  use_nodes = NULL
)
}
\arguments{
\item{in_graph}{the cc_graph object}

\item{in_assign}{the colors generated by \code{assign_colors}}

\item{node_communities}{the communities generated by \code{label_communities}}

\item{use_nodes}{the list of nodes to actually use}
}
\value{
list
}
\description{
takes a \code{cc_graph} object and transforms it into something that can
be visualized using \code{visNetwork}
}
