% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\docType{class}
\name{gsea_features-class}
\alias{gsea_features-class}
\title{GSEA feature class}
\description{
class to hold features undergoing GSEA
}
\section{Slots}{

\describe{
\item{\code{ranks}}{a named vector of ranks}

\item{\code{annotation}}{annotation object}
}}

