% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea_features}
\alias{gsea_features}
\title{gsea_features constructor}
\usage{
gsea_features(ranks, annotation)
}
\arguments{
\item{ranks}{a named vector of ranks}

\item{annotation}{the annotation object}
}
\value{
gsea_features-class
}
\description{
Creates a gsea_features-class object.
}
