% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{gsea_result-class}
\alias{gsea_result-class}
\alias{gsea_result}
\title{the gsea results class}
\description{
the gsea results class
}
\section{Slots}{

\describe{
\item{\code{ranks}}{the ranked features}
}}

