% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea_result_to_fgsea}
\alias{gsea_result_to_fgsea}
\title{convert gsea object}
\usage{
gsea_result_to_fgsea(in_gsea)
}
\arguments{
\item{in_gsea}{the gsea_result object}
}
\value{
data.table
}
\description{
Takes an `gsea_result`, and converts it to the table expected by `fgsea`.
This should only be done on those that have `gsea` as the *Enrichment Method*.
}
