% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{have_parallel}
\alias{have_parallel}
\title{choose apply}
\usage{
have_parallel()
}
\arguments{
\item{x}{the list object to work on}

\item{fun}{the function to be called}

\item{...}{other parameters}
}
\value{
list results
}
\description{
based on whether the parallel package is loaded and options(mc.cores) is not 1,
passes data on to the base lapply or the mclapply function, and returns
the results.
}

