% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{hypergeometric_feature_enrichment}
\alias{hypergeometric_feature_enrichment}
\title{do hypergeometric enrichment}
\usage{
hypergeometric_feature_enrichment(
  hypergeometric_features,
  direction = "over",
  p_adjust = "BH",
  min_features = 1
)
}
\arguments{
\item{hypergeometric_features}{a hypergeometric_features object}

\item{direction}{which direction to do the enrichment (over or under)}

\item{p_adjust}{how to correct the p-values (default is "BH")}

\item{min_features}{how many features should be annotated before testing it?}
}
\value{
enriched_result
}
\description{
do hypergeometric enrichment
}
\details{
The \code{min_features} argument here applies to the minumum number of features an annotation has
  from the universe of features supplied, \bold{not} the minumum number of features from the differential
  list.
  For more about the p-value adjustment, see \code{stats::p.adjust}
}
