% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{hypergeometric_features}
\alias{hypergeometric_features}
\title{hypergeometric_features constructor}
\usage{
hypergeometric_features(significant, universe, annotation)
}
\arguments{
\item{significant}{the features that are significant}

\item{universe}{the background or universe of features measured}

\item{annotation}{the annotation object}
}
\value{
hypergeometric_features-class
}
\description{
function to easily construct a hypergeometric_features-class object.
}
