% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{hypergeometric_result}
\alias{hypergeometric_result}
\title{the hypergeometric results class}
\usage{
hypergeometric_result(features, universe, annotation, statistics)
}
\description{
the hypergeometric results class
}
\section{Slots}{

\describe{
\item{\code{features}}{the significant features}

\item{\code{universe}}{the universe of features}
}}

