% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_executables}
\alias{install_executables}
\title{install executables}
\usage{
install_executables(path = "~/bin")
}
\arguments{
\item{path}{the path to put the executable scripts}
}
\value{
the listing of the files.
}
\description{
move executables to user location, default is ~/bin and changes their
permissions to make them executable.
}
