% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_annotations.R
\name{json_2_annotation}
\alias{json_2_annotation}
\title{json to annotation}
\usage{
json_2_annotation(json_file)
}
\arguments{
\item{json_file}{the json annotation file}
}
\value{
annotation object
}
\description{
Given a JSON based annotation object, read it in and create the `annotation`
for actually doing enrichment.
}
