% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_annotations.R
\name{json_annotation_reversal}
\alias{json_annotation_reversal}
\title{annotation reversal}
\usage{
json_annotation_reversal(
  json_file,
  out_file = "annotations.json",
  feature_type = NULL,
  annotation_type = NULL
)
}
\arguments{
\item{json_file}{the json file to use}

\item{out_file}{the json file to write out to}

\item{feature_type}{the type of features}

\item{annotation_type}{the type of annotations}
}
\value{
the json object, invisibly
}
\description{
Given a JSON file of features to annotations, reverse to turn it into
annotations to features, and optionally add some meta-information about them.
}
