% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{kable_annotation_table}
\alias{kable_annotation_table}
\title{print table kable}
\usage{
kable_annotation_table(annotation_gene_table, header_level = 3, cat = TRUE)
}
\arguments{
\item{annotation_gene_table}{list of tables}

\item{header_level}{what header level should the labels be done at?}

\item{cat}{whether to write it directly, or just return the table for later}
}
\value{
character
}
\description{
print the annotation gene table in \code{knitr::kable} format
}
