% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{label_go_communities}
\alias{label_go_communities}
\title{label GO communities}
\usage{
label_go_communities(community_defs)
}
\arguments{
\item{community_defs}{list of GO communities (from \code{assign_communities})}
}
\value{
list
}
\description{
determine the labels of Gene Ontology communities by deciding on the most generic
GO terms in each community.
}

