% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{name_go_communities}
\alias{name_go_communities}
\title{name GO communities}
\usage{
name_go_communities(community_defs)
}
\arguments{
\item{community_defs}{list of GO communities (from \code{assign_communities})}
}
\value{
list
}
\description{
determine the names of Gene Ontology communities by deciding on the most generic
GO terms in each community.
}

