% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/graph_visualization.R
\name{remove_edges}
\alias{remove_edges}
\alias{remove_edges,character,numeric-method}
\alias{remove_edges,cc_graph,numeric-method}
\title{remove edges}
\usage{
remove_edges(edge_obj, cutoff, edge_attr = "weight", value_direction = "under")

\S4method{remove_edges}{character,numeric}(edge_obj, cutoff, edge_attr = "weight", value_direction = "under")

\S4method{remove_edges}{cc_graph,numeric}(edge_obj, cutoff, edge_attr = "weight", value_direction = "under")
}
\arguments{
\item{edge_obj}{cc_graph}

\item{cutoff}{the cutoff to use}

\item{edge_attr}{which attribute to use}

\item{value_direction}{remove edges with value under or over}
}
\value{
nothing

cc_graph
}
\description{
given a RCy3 network connection, remove edges according to provided
values.
}
