% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{show,binomial_result-method}
\alias{show,binomial_result-method}
\title{show binomial_result}
\usage{
\S4method{show}{binomial_result}(object)
}
\arguments{
\item{object}{the binomial_result object to show}
}
\description{
show binomial_result
}
