% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{methods}
\name{show,combined_statistics-method}
\alias{show,combined_statistics-method}
\title{show combined_statistics}
\usage{
\S4method{show}{combined_statistics}(object)
}
\arguments{
\item{object}{\linkS4class{combined_statistics}}
}
\description{
show combined_statistics
}

