% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{show,enriched_result-method}
\alias{show,enriched_result-method}
\title{show enriched_result}
\usage{
\S4method{show}{enriched_result}(object)
}
\arguments{
\item{object}{the enriched_result object to show}
}
\description{
show enriched_result
}
