% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{methods}
\name{show,node_assign-method}
\alias{show,node_assign-method}
\title{show node_assign}
\usage{
\S4method{show}{node_assign}(object)
}
\arguments{
\item{object}{the node_assign to see}
}
\description{
show node_assign
}

