% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{methods}
\name{show,significant_annotations-method}
\alias{show,significant_annotations-method}
\title{show signficant_annotations}
\usage{
\S4method{show}{significant_annotations}(object)
}
\description{
show signficant_annotations
}

