% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{significant_annotations}
\alias{significant_annotations}
\title{significant annotations}
\usage{
significant_annotations(significant, measured, sig_calls = NULL)

significant_annotations(significant, measured, sig_calls = NULL)
}
\arguments{
\item{significant}{logical matrix of annotations (rows) and experiments (columns)}

\item{measured}{logical matrix of annotations (rows) and experiments (columns)}

\item{sig_calls}{character vector of deparsed calls that resulted in signficant and
measured}
}
\description{
The \code{significant_annotations} class holds which annotations from which
enrichment were both \strong{measured} and \strong{significant}. Each of these
slots is a \emph{logical matrix} with rows named by \emph{annotation_id} and
columns named by the names of the \code{\link{enriched_result}} that was combined.

Makes a new significant_annotation while checking that everything is valid.
}
\section{Slots}{

\describe{
\item{\code{significant}}{logical matrix}

\item{\code{measured}}{logical matrix}

\item{\code{sig_calls}}{character representations of calls used to filter the data}
}}

