% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{table_from_graph}
\alias{table_from_graph}
\title{table from graph}
\usage{
table_from_graph(in_graph, in_assign = NULL, community_info = NULL)
}
\arguments{
\item{in_graph}{the \code{cc_graph} object}

\item{in_assign}{the \code{node_assign} object}

\item{community_info}{the \code{community_info} object}
}
\value{
data.frame
}
\description{
Creates a table from the annotation graph, and if provided, adds the
community information to the table.
}

