% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{vis_visnetwork}
\alias{vis_visnetwork}
\title{vis in visNetwork}
\usage{
vis_visnetwork(in_graph_info)
}
\arguments{
\item{in_graph_info}{the graph structure from \code{graph_to_visnetwork}}
}
\description{
Visualize a \code{cc_graph} in \code{visNetwork}, with selection for communities
if that exists.
}
