% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_classes.R
\name{cc_graph}
\alias{cc_graph}
\title{cc_graph}
\usage{
cc_graph(graph, significant)

cc_graph(graph, significant)
}
\arguments{
\item{graph}{the \linkS4class{graphNEL}}

\item{significant}{a matrix indicating which nodes are significant in which experiment}
}
\description{
A \code{cc_graph} class is a \code{graphNEL} with the added slot of
\code{significant}, a matrix of rows (nodes / annotations) and whether
they were found to be significant in a given enrichment (columns). This
matrix is used for classifying the annotations into different groups, and
generating either pie-charts or coloring the nodes in a visualization.

constructs a \emph{cc_graph} given a \linkS4class{graphNEL} and a \emph{significant} matrix.
}
\section{Slots}{

\describe{
\item{\code{significant}}{numeric matrix of ones and zeros}
}}

