% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant_results_calls.R
\name{combined_significant_calls}
\alias{combined_significant_calls}
\title{get significant annotations calls}
\usage{
combined_significant_calls(in_results, queries)
}
\arguments{
\item{in_results}{a \linkS4class{combined_enrichment} object}

\item{queries}{a list of queries that can form a call object}
}
\value{
\linkS4class{combined_enrichment} object
}
\description{
In the case where we have a \linkS4class{combined_enrichment} and we want
to get all of the significant annotations from each of them, and put them
together so we can start doing real meta-analysis.
}
\details{
Note that this function returns the original \linkS4class{combined_enrichment} object with a modified
\linkS4class{combined_statistics} slot where the significant annotations have been added in.
}
