% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{enriched_result-class}
\alias{enriched_result-class}
\alias{enriched_result}
\alias{hypergeometric_result}
\alias{hypergeometric_result-class}
\alias{gsea_result-class}
\alias{gsea_result}
\alias{show,enriched_result-method}
\alias{binomial_result-class}
\alias{binomial_result}
\alias{show,binomial_result-method}
\title{the enriched results class}
\usage{
hypergeometric_result(features, universe, annotation, statistics)

\S4method{show}{enriched_result}(object)

hypergeometric_result(features, universe, annotation, statistics)

\S4method{show}{binomial_result}(object)
}
\arguments{
\item{features}{the features that were differentially expressed (see details)}

\item{universe}{all of the features that were measured}

\item{annotation}{an \code{\link{annotation}} object}

\item{statistics}{a \code{\link{statistical_results}} object}

\item{object}{the binomial_result object to show}
}
\value{
hypergeometric_result
}
\description{
given all the slots for an \code{\link{enriched_result}}, checks that all
the data is self-consistent, and creates the \code{\link{enriched_result}} object.
}
\section{Slots}{

\describe{
\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \code{\link{statistical_results}} object}

\item{\code{features}}{the significant features}

\item{\code{universe}}{the universe of features}

\item{\code{ranks}}{the ranked features}

\item{\code{positivefc}}{the positive log-fold-changed genes, a vector of class ANY}

\item{\code{negativefc}}{the negative log-fold-changed genes}
}}

