% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{hypergeometric_result}
\alias{hypergeometric_result}
\title{creates hypergeometric result}
\usage{
hypergeometric_result(features, universe, annotation, statistics)
}
\arguments{
\item{features}{the features that were differentially expressed (see details)}

\item{universe}{all of the features that were measured}

\item{annotation}{an \code{\link{annotation}} object}

\item{statistics}{a \code{\link{statistical_results}} object}
}
\value{
hypergeometric_result
}
\description{
given all the slots for an \code{\link{enriched_result}}, checks that all
the data is self-consistent, and creates the \code{\link{enriched_result}} object.
}
