% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_classes.R
\name{enriched_result}
\alias{enriched_result}
\title{the enriched results class}
\usage{
enriched_result(features, universe, annotation, statistics)
}
\description{
the enriched results class
}
\section{Slots}{

\describe{
\item{\code{features}}{the "features" of interest, a vector of class ANY}

\item{\code{universe}}{all of the "features" in the background}

\item{\code{annotation}}{list giving the annotation to feature relationship}

\item{\code{statistics}}{a \linkS4class{statistical_results} object}
}}

