% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_enrichments.R
\name{extract_enrich_stats}
\alias{extract_enrich_stats}
\title{extract enrich stats}
\usage{
extract_enrich_stats(enrichment_result)
}
\arguments{
\item{enrichment_result}{the enrichment result object}
}
\value{
data.frame
}
\description{
Extract statistical table from a single enrichment object.
}
