% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocats.R
\name{gocats_to_annotation}
\alias{gocats_to_annotation}
\title{gocats to annnotations}
\usage{
gocats_to_annotation(
  ancestors_file = "ancestors.json",
  namespace_file = "namespace.json",
  annotation_type = "gocatsGO",
  feature_type = "Uniprot"
)
}
\arguments{
\item{ancestors_file}{the ancestors.json file from gocats (required)}

\item{namespace_file}{the namespace.json file from gocats (optional)}

\item{annotation_type}{what annotations are we making? (gocatsGO by default)}

\item{feature_type}{what type of features are we using (assume Uniprot)}
}
\value{
annotation object
}
\description{
Transforms a gocats ancestors JSON list to a GO annotation object.
}
