% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gocats.R
\name{gocats_to_annotation}
\alias{gocats_to_annotation}
\title{gocats to annnotations}
\usage{
gocats_to_annotation(
  ancestors_file = "ancestors.json",
  namespace_file = "namespace.json",
  annotation_type = "gocatsGO",
  feature_type = "Uniprot",
  feature_translation = NULL,
  feature_min = 5,
  feature_max = 5000
)
}
\arguments{
\item{ancestors_file}{the ancestors.json file from gocats (required)}

\item{namespace_file}{the namespace.json file from gocats (optional)}

\item{annotation_type}{what annotations are we making? (gocatsGO by default)}

\item{feature_type}{what type of features are we using (assume Uniprot)}

\item{feature_translation}{a data.frame used to convert the feature IDs}

\item{feature_min}{minimum number of features annotated (default: 5)}

\item{feature_max}{maximum number of features annotated (default: 5000)}
}
\value{
annotation object
}
\description{
Transforms a gocats ancestors JSON list to a GO annotation object.
}
