% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_enrichment.R
\name{hypergeometric_feature_enrichment}
\alias{hypergeometric_feature_enrichment}
\title{do hypergeometric enrichment}
\usage{
hypergeometric_feature_enrichment(hypergeom_features, direction = "over",
  p_adjust = "none")
}
\arguments{
\item{hypergeom_features}{a hypergeometric_features object}

\item{direction}{which direction to do the enrichment (over or under)}

\item{p_adjust}{how to correct the p-values (default is "none")}
}
\value{
enriched_result
}
\description{
do hypergeometric enrichment
}
