% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_visualization.R
\name{label_communities}
\alias{label_communities}
\title{label communities}
\usage{
label_communities(community_defs)
}
\arguments{
\item{community_defs}{the communities from \code{assign_communities}}
}
\value{
list
}
\description{
label communities
}

