% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\name{show,significant_annotations-method}
\alias{show,significant_annotations-method}
\title{show signficant_annotations}
\usage{
\S4method{show}{significant_annotations}(object)
}
\arguments{
\item{object}{the significant annotations object to show}
}
\description{
show signficant_annotations
}
